requirements.txt
asgiref==3.5.2
autopep8==1.7.0
click==8.1.3
fastapi==0.68.2
h11==0.14.0
motor==3.0.0
pycodestyle==2.9.1
pydantic==1.10.2
pymongo==4.2.0
python-dotenv==0.21.0
starlette==0.14.2
toml==0.10.2
typing_extensions==4.4.0
uvicorn==0.15.0
